Remove trailing whitespace from file license headers
Since presumably these licensing headers are never going to be touched in the normal course of development, the trailing whitespace here won't be cleaned up by simply waiting for some developer to touch this area of the code and automatically format it along with formatting his submission.

If a contributor is not using automatic formatting which cleans up trailing whitespace, then one way to avoid trailing whitespace issues is to make it visible in one's editor. In Visual Studio use Edit > Advanced > View White Space or use ctrl-r, ctrl-w. In vim see :help listchars.